Improve/simplify camera transition code
Now the Camera keeps track of the transition data, and the Tracker only
signals that the target position/zoom factor has been changed.
Ths fixes the issue with paused cameras causing jumps in the transition.

Also fix issues with Luadoc (of related code) and cover more functions
and types with documentation.